pyproject.toml
[project]
name = "vanilla-nn-price-prediction"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.8.0",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
    "python-dotenv>=1.1.1",
    "datasets==3.6.0",
    "pydantic>=2.0.0",
    "pandas>=2.3.3",
    "plotly>=6.3.0",
]